Refine the settings of dependencies
