+++ package.json
@@ -47,6 +47,8 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "observable-hooks": "^3.1.2",
+    "rxjs": "^6.6.2",
     "react-spring": "^8.0.18",
     "rimraf": "^3.0.2",
     "styled-components": "^5.1.1",
@@ -54,8 +56,7 @@
     "ts-loader": "^5.3.3",
     "typescript": "^3.9.7",
     "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0",
-    "observable-hooks": "^3.1.2"
+    "webpack-cli": "^3.3.0"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org"
