+++ packages/core/package.json
@@ -34,10 +34,8 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "rxjs": "^6.6.2",
     "observable-hooks": "^3.1.2"
-  },
-  "dependencies": {
-    "rxjs": "^6.6.0"
   },
   "bugs": {
     "url": "https://github.com/garfieldduck/intrasections/issues"
