+++ packages/core/src/types/IntersectionObserverConfig.ts
@@ -1,3 +1,10 @@
+export interface Margin {
+  top: number;
+  right: number;
+  bottom: number;
+  left: number;
+}
+
 export interface IntersectionObserverConfig {
   /**
    * Threshold at which to trigger callback.
