+++ packages/plot/src/components/StickyPlot.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useMemo } from 'react';
+import React, { useRef } from 'react';
 import {
   // utils
   getStickyPosition,
