+++ packages/trigger/package.json
@@ -37,6 +37,9 @@
     "react-dom": "^16.13.1",
     "@react-scrolly/core": "^0.2.1-alpha.4"
   },
+  "devDependencies": {
+    "@react-scrolly/core": "^0.2.1-alpha.4"
+  },
   "bugs": {
     "url": "https://github.com/garfieldduck/intrasections/issues"
   },
